add route to verify a user by id, protect that by role, refactor admin page a lil, add beginings of VerifyUsers table
