+++ packages/frontend/src/components/UserStatusTable.tsx
@@ -14,34 +14,30 @@ import ThumbDownIcon from '@mui/icons-material/ThumbDown';
 import { useRecoilValue } from 'recoil';
 import { CurrentUserSignupStatus } from '../state/store';
 
+const generateTableRow = (criteria: string, status: boolean, help: string) => (
+  <TableRow
+    key={criteria}
+    sx={{ '&:last-child td, &:last-child th': { border: 0 } }}
+  >
+    <TableCell component="th" scope="row">
+      {criteria}
+    </TableCell>
+    <TableCell component="th" scope="row">
+      {status ? (
+        <ThumbUpIcon color="success" />
+      ) : (
+        <ThumbDownIcon color="error" />
+      )}
+    </TableCell>
+    <TableCell>
+      <Typography>{help}</Typography>
+      <IconButton />
+    </TableCell>
+  </TableRow>
+);
+
 function UserStatusTable() {
   const signupStatus = useRecoilValue(CurrentUserSignupStatus);
-
-  const generateTableRow = (
-    criteria: string,
-    status: boolean,
-    help: string,
-  ) => (
-    <TableRow
-      key={criteria}
-      sx={{ '&:last-child td, &:last-child th': { border: 0 } }}
-    >
-      <TableCell component="th" scope="row">
-        {criteria}
-      </TableCell>
-      <TableCell component="th" scope="row">
-        {status ? (
-          <ThumbUpIcon color="success" />
-        ) : (
-          <ThumbDownIcon color="error" />
-        )}
-      </TableCell>
-      <TableCell>
-        <Typography>{help}</Typography>
-        <IconButton />
-      </TableCell>
-    </TableRow>
-  );
 
   return (
     <TableContainer>
